requirements.txt
numpy
scipy
librosa
numba
tqdm
hydra-core
pyloudnorm
tensorboard
textgrid
speechbrain